validateForm function in flux actions
